requirements.txt
flask== 2.3.2
gunicorn==19.9.0
line-bot-sdk==2.4.2
numpy==1.23.2 
beautifulsoup4==4.6.3 
requests==2.20.1 
scipy 
pandas==1.2.4 
lxml==4.5.2 
line-bot-sdk==1.19.0 
datetime  
matplotlib 
gspread==4.0.1 
oauth2client==4.1.3 
pyimgur 
mpl_finance 
arrow